Attempt to make onig_match and onig_search take const regex_t*.
Believing `onig_match` and `onig_search` to be thread-safe with respect to the
regular expression struct, this is an attempt to make those functions take a
const-pointer to the regular expression struct for illustration purposes.

This commit will break the build because these two functions end up calling
`set_bm_backward_skip`, which modifies the regular expression struct in a
potentially thread unsafe way.